Reverted customizable formatting. See note.
It's still an option but not default.
Sometimes adding more configuration and filters to the logging config is very hard.

Sometimes even impossible. For example:
To make it work with Log4j2 you need to use marker pattern selector. But when you use that, you cannot use other selectors...
You can always write your own custom pattern selector but that's tiresome and can be easily avoided with default formatting.
So unless someone needs to change the default formatting the configuration is much easier.